Add logic for Vault
Add a new package called vault which handles the API client generation
and the logic to read a secret from Vault. If the secret was found in
Vault the data of this secret is returned in the format of the data
field in a Kubernetes Secret.

Replace the newPodForCR function with a newSecretForCR function, which
returns a new Kubernetes secret.

Adjust the Custom Resource Definition. The CR requires a path and a
type. The path is the path to a Vault secret and the type is the type of
the Kubernetes secret which should be created.