Improve Operator logic
Add a field named 'keys' to the CRD. The keys field is used to only
include the provided keys from a Vault secret in the corresponding
Kubernetes secret. If the field is omitted all keys from the Vault
secret will be appear in the Kubernetes secret.

Create the Vault API client during the start of the Operator. When the
API client was successfully created, a RenewToken function is called to
renew the provided Vault token. For this we introduced a new environment
variable 'VAULT_TOKEN_TTL' with the TTL of the token. After the half of
the TTL time is over the token will be renewed.

If a corresponding secret for a CR already exists we update the existing
secret.