Add Makefile
Add Makefile to build and release the Docker image for the Operator.
During the build process we pass some build information to the go build
command via ldflags.

Because of the operator-sdk build command we can only populate one
variable. So we need to adjust the version handling to get clean values
for the build information.