Prepare Helm chart for Kubernetes Auth Method
If the Kubernetes Auth Method is set in the vaules of the Helm chart, we
create the necessary rules for the service account of the Vault Secrets
Operator.

Add values, which populates the needed environment variables in the
deployment of the Operator.

Fix an error in the cluster role template for the creation of the
service monitor for Prometheus.