Implement KV Secrets Engine - Version 2
Add implementation for the KV2 secrets engine. In the custom resource
it's now possible to specify the secret engine under which the secret
exists. The 'secretEngine' field in the CR must be 'kv1' or 'kv2'. To
not introduce a breaking change the field can be omitted then the KV1
secrets engine will be used.

If the the KV2 secret engine is used we will check the path if it
contains 'data' at the second position (first position is the path under
with the secret engine was enabled). If the path does not contain 'data'
we will add it.

The retuned secret from a KV2 secrets engine has another structure then
one from a KV1 secret engine. So we try to parse the 'data' field in the
returned secret data (there is also a 'metadata' field). Because of
behaviour we also check now the resulting data for the Kubernetes
secret. If the length of this data is 0 we return an error instead of
creating a secret without any data.