Automatic determination of KV version
We use the logic from the Vault cli to determine the used version of the
secrets engine. This also fixes the problem, when a KVv2 secrets engine
was mounted under a nested path.

The secretEngine field in the CR is no longer needed, but stays. So we
do not introduce a breaking change and we can reuse it later if we
implement other secret engines.

TODO:
- Improve description for the secretEngine field
- Improve comments