Add logic for automatic reconciliation
If an environment variable VAULT_RECONCILIATION_TIME is specified and
greater than zero, we set the reconciliation time to the value of this
variable in seconds.

When the reconciliation time is greater than zero we reconcil ever vault
secret which does not specify a version in the CR after this time.

With this change it is possible to automatic update a Kubernetes secret
if the secret in Vault changes.