Add vaule to set deployment strategy to Helm chart
It is now possible to specify the deployment strategy which should be
used for the deployment of the operator.

Currently there seems to be a bug in the operator, which cause problems
during an update. Therefor it is recommended to set the strategy type to
recreate if a replication count of one is used:

deploymentStrategy:
  type: Recreate