Restrict the controller to a particular namespaces
We are using the NewCache option for the manager to watch particular
namespaces. If a comma seperated list of namespaces is passed to the
operator via the WATCH_NAMESPACE environment variable only these
namespaces will be watched for CRs. If the WATCH_NAMESPACE variable is
empty all namespaces will be watched.