Parameterise token renewal intervals
For particular Vault setups, the default values for token renewals might
not be a good fit. This commit adds a parametrisation, so that the time
between a successful or failed token renewal and the next renewal attempt
can be controlled via optional environment variables.

VAULT_TOKEN_RENEWAL_INTERVAL:
    The time (in seconds) between a successful token renewal and the
    next renewal attempt. Defaults to 50% of the token lease duration.

VAULT_TOKEN_RENEWAL_RETRY_INTERVAL:
    The time (in seconds) between a failed token renewal and the next
    renewal attempt. Defaults to 30 seconds.

Refs #48