Fix RBAC generation
The created RBAC manifests were not in sync with the specified RBAC
markers. We added the RBAC markers for the ConfigMaps and Events and
also for the finalizers. This is needed to fix the bug mentioned in #65.

Now we can generate the RBAC rules via make manifests again. Then we can
use the generated file from confgi/rbac/role.yaml in the Helm chart.