Use probes to check renew token status
When the renewal of the Vault token fails to often, the liveness and
readiness probe of the operator will fail. For this we are adding a
counter, which is increased everytime the renewal of the Vault token
fails.

When the counter if greater or equal then 5 the readiness probe will
fail and if it is greater or equal then 10 the liveness probe will fail.