Remove created-by label
This commit removes the "created-by" label, which was added to the
Kubernetes secrets created by the Vault Secrets Operator. Since this
label wasn't then used for any further actions it should be save to
remove.

If a user depends on the label for his workflows it can be added as
label to the VaultSecret and will then also be added to the Kubernetes
secret.