Restrict Operator to Vault Namespace (#246)
It is now possible to restrict the operator to a specific Vault
Namespace. When the `VAULT_RESTRICT_NAMESPACE` environment variable is
set to `true` the operator only reconciles secrets where the
`spec.vaultNamespace` field is the same as the `VAULT_NAMESPACE`
environment variable. VaultSecrets with another `spec.vaultNamespace`
value will be ignored.

When this feature is enabled the operator can not be used with nested
Vault Namespaces.